Splash is DeepLinkView
