+++ app/src/main/AndroidManifest.xml
@@ -26,9 +26,6 @@
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-
-        <activity android:name="com.markus.subscity.ui.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
 
@@ -47,6 +44,8 @@
 
             </intent-filter>
         </activity>
+
+        <activity android:name="com.markus.subscity.ui.main.MainActivity" />
 
         <activity
             android:name="com.markus.subscity.ui.movie.MovieActivity"
