+++ app/src/main/java/com/markus/subscity/ui/splash/SplashActivity.kt
@@ -5,31 +5,74 @@ import com.arellomobile.mvp.MvpAppCompatActivity
 import com.arellomobile.mvp.presenter.InjectPresenter
 import com.arellomobile.mvp.presenter.ProvidePresenter
 import com.markus.subscity.dagger.SubsCityDagger
+import com.markus.subscity.extensions.analytics
+import com.markus.subscity.ui.cinema.CinemaActivity
 import com.markus.subscity.ui.city.CityActivity
+import com.markus.subscity.ui.deeplink.DeepLinkPresenter
+import com.markus.subscity.ui.deeplink.DeepLinkView
+import com.markus.subscity.ui.deeplink.isFromDeepLink
 import com.markus.subscity.ui.main.MainActivity
+import com.markus.subscity.ui.movie.MovieActivity
 
 /**
  * @author Vitaliy Markus
  */
-class SplashActivity : MvpAppCompatActivity(), SplashView {
-
+class SplashActivity : MvpAppCompatActivity(), SplashView, DeepLinkView {
 
     @InjectPresenter
     lateinit var splashPresenter: SplashPresenter
 
+    @InjectPresenter
+    lateinit var deepLinkPresenter: DeepLinkPresenter
 
     @ProvidePresenter
     fun splashPresenter(): SplashPresenter {
         return SubsCityDagger.component.createSplashPresenter()
     }
 
+    @ProvidePresenter
+    fun deepLinkPresenter(): DeepLinkPresenter {
+        return SubsCityDagger.component.createDeepLinkPresenter()
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        splashPresenter.checkFirstLaunch()
+
+        if (isFromDeepLink) {
+            deepLinkPresenter.performDeepLink(intent.data)
+        } else {
+            splashPresenter.checkFirstLaunch()
+        }
     }
 
     override fun showMain() {
         MainActivity.start(this)
+        finish()
+    }
+
+    override fun showMovies() {
+        MainActivity.start(this, MainActivity.Companion.Mode.MOVIES)
+        analytics().logOpenMovies(true)
+        finish()
+    }
+
+    override fun showCinemas() {
+        MainActivity.start(this, MainActivity.Companion.Mode.CINEMAS)
+        analytics().logOpenCinemas(true)
+        finish()
+    }
+
+    override fun showMovie(movieId: Long) {
+        analytics().logOpenMovie(movieId, null, true)
+        MainActivity.start(this, MainActivity.Companion.Mode.MOVIES)
+        MovieActivity.start(this, movieId)
+        finish()
+    }
+
+    override fun showCinema(cinemaId: Long) {
+        analytics().logOpenCinema(cinemaId, null, true)
+        MainActivity.start(this, MainActivity.Companion.Mode.CINEMAS)
+        CinemaActivity.start(this, cinemaId)
         finish()
     }
