+++ app/src/main/java/com/markus/subscity/ui/splash/SplashView.kt
@@ -1,13 +1,13 @@
 package com.markus.subscity.ui.splash
 
 import com.arellomobile.mvp.MvpView
-import com.arellomobile.mvp.viewstate.strategy.SingleStateStrategy
+import com.arellomobile.mvp.viewstate.strategy.OneExecutionStateStrategy
 import com.arellomobile.mvp.viewstate.strategy.StateStrategyType
 
 /**
  * @author Vitaliy Markus
  */
-@StateStrategyType(SingleStateStrategy::class)
+@StateStrategyType(OneExecutionStateStrategy::class)
 interface SplashView : MvpView {
 
     fun showMain()
